Convert the getData() funciton into onPageload  class
